+++ SAKPaaS/src/app/app.component.ts
@@ -6,7 +6,7 @@ import { TranslateService } from '@ngx-translate/core';
 import { CookieProviderService } from 'src/app/core/services/cookie-provider.service';
 import { MixpanelService, MixpanelId } from './core/services/mixpanel.service';
 import { DOCUMENT } from '@angular/common';
-import { PwaRequestPromptService } from './generated/services/pwa-request-prompt.service';
+import { PwaRequestPromptService } from './core/services/pwa-request-prompt.service';
 
 @Component({
   selector: 'app-root',
